requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
python-dateutil==2.9.0.post0
